pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safe-codegen"
version = "0.1.0"
description = "Multi-agent LangGraph-based safe code generation pipeline with layered defenses and progressive caching."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
dependencies = [
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "langchain-openai>=0.1.0",
  "langchain-community>=0.3.0",
  "openai>=1.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "python-dotenv>=1.0.0",
  "sqlalchemy>=2.0.0",
  "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.23.0",
  "autogen-agentchat>=0.4.0",
  "autogen-ext[openai]>=0.4.0",
]

[project.scripts]
safe-codegen = "safe_codegen.runner:main"

[tool.setuptools.packages.find]
where = ["src"]

